refactor:#104 GNB가 MenuBar를 대체
