--- src/components/layouts/menu-bar.tsx
@@ -1,11 +1,0 @@
-import GlobalNavigationBar from '@/components/commons/global-navigation-bar';
-
-const MenuBar = () => {
-  return (
-    <div className="fixed bottom-4 left-4 right-4">
-      <GlobalNavigationBar />
-    </div>
-  );
-};
-
-export default MenuBar;
